Create abstract superclass
