+++ lib/fs-token-persistence.spec.js
@@ -4,7 +4,7 @@ const fs = require('fs')
 const tmp = require('tmp')
 const readFile = require('fs-readfile-promise')
 const { expect } = require('chai')
-const TokenPersistence = require('./token-persistence')
+const FsTokenPersistence = require('./fs-token-persistence')
 const githubAuth = require('./github-auth')
 
 describe('Token persistence', function() {
@@ -14,7 +14,7 @@ describe('Token persistence', function() {
   let path, persistence
   beforeEach(function() {
     path = tmp.tmpNameSync()
-    persistence = new TokenPersistence({ path })
+    persistence = new FsTokenPersistence({ path })
   })
 
   context('when the file does not exist', function() {
