+++ lib/token-persistence.js
@@ -1,53 +1,38 @@
 'use strict'
 
-const fsos = require('fsos')
-const githubAuth = require('./github-auth')
 const log = require('./log')
 
 // This is currently bound to the legacy github auth code. That will be
 // replaced with a dependency-injected token provider.
 class TokenPersistence {
-  constructor({ path }) {
-    this.path = path
-
+  constructor() {
     this.noteTokenAdded = this.noteTokenAdded.bind(this)
     this.noteTokenRemoved = this.noteTokenRemoved.bind(this)
   }
 
   async initialize() {
-    let contents
-    try {
-      contents = await fsos.get(this.path)
-    } catch (e) {
-      if (e.code === 'ENOENT') {
-        contents = '[]'
-      } else {
-        throw e
-      }
-    }
-
-    const tokens = JSON.parse(contents)
-    tokens.forEach(tokenString => {
-      githubAuth.addGithubToken(tokenString)
-    })
+    throw Error('initialize() is not implemented')
   }
 
-  async save() {
-    const tokens = githubAuth.getAllTokenIds()
-    await fsos.set(this.path, JSON.stringify(tokens))
+  async onTokenAdded(token) {
+    throw Error('onTokenAdded() is not implemented')
   }
 
   async noteTokenAdded(token) {
     try {
-      await this.save()
+      await this.onTokenAdded(token)
     } catch (e) {
       log.error(e)
     }
   }
 
+  async onTokenRemoved(token) {
+    throw Error('onTokenRemoved() is not implemented')
+  }
+
   async noteTokenRemoved(token) {
     try {
-      await this.save()
+      await this.onTokenRemoved(token)
     } catch (e) {
       log.error(e)
     }
